Refactor LLM into backends
