+++ consultation_analyser/pipeline/ml_pipeline.py
@@ -1,7 +1,6 @@
 import logging
 from uuid import UUID
 
-
 from consultation_analyser.consultations import models
 
 from .backends.topic_backend import TopicBackend
