+++ consultation_analyser/pipeline/processing.py
@@ -1,5 +1,6 @@
 from consultation_analyser.hosting_environment import HostingEnvironment
 from consultation_analyser.pipeline.backends.bertopic import BERTopicBackend
+from consultation_analyser.pipeline.backends.dummy_llm_backend import DummyLLMBackend
 from consultation_analyser.pipeline.batch_calls import BatchJobHandler
 from consultation_analyser.pipeline.llm_summariser import (
     create_llm_summaries_for_consultation,
@@ -7,16 +8,15 @@ from consultation_analyser.pipeline.llm_summariser import (
 from consultation_analyser.pipeline.ml_pipeline import save_themes_for_consultation
 
 
-def process_consultation_themes(consultation, topic_backend=None, llm=None):
+def process_consultation_themes(consultation, topic_backend=None, llm_backend=None):
     if not topic_backend:
         topic_backend = BERTopicBackend()
 
-    save_themes_for_consultation(consultation.id, topic_backend)
+    if not llm_backend:
+        llm_backend = DummyLLMBackend()
 
-    if llm:
-        create_llm_summaries_for_consultation(consultation, llm)
-    else:
-        create_llm_summaries_for_consultation(consultation)
+    save_themes_for_consultation(consultation.id, topic_backend)
+    create_llm_summaries_for_consultation(consultation, llm_backend)
 
 
 def run_processing_pipeline(consultation):
