+++ tests/integration/test_llm_summariser.py
@@ -1,6 +1,7 @@
 import pytest
 
 from consultation_analyser import factories
+from consultation_analyser.pipeline.backends.dummy_llm_backend import DummyLLMBackend
 from consultation_analyser.pipeline.llm_summariser import create_llm_summaries_for_consultation
 
 
@@ -16,7 +17,7 @@ def test_create_llm_summaries_for_consultation():
     assert not theme.summary
     assert not theme.short_description
 
-    create_llm_summaries_for_consultation(consultation)
+    create_llm_summaries_for_consultation(consultation, DummyLLMBackend())
 
     theme.refresh_from_db()
     assert theme.summary
